2024-07-01 KJY, 통계화면 에러 수정 - 통계화면 에러 수정 - 로그인 화면 room db 수정

+++ app/src/main/java/com/example/cse_study_and_learn_application/model/StatisticsModels.kt
@@ -9,4 +9,4 @@ package com.example.cse_study_and_learn_application.model
  * @author kjy
  * @since 2024-03-24
  */
-data class CategorySuccessRatio(val title: String, val ratio: Int)
+data class CategorySuccessRatio(val title: String, val ratio: Double)
